Update rentpath/clj-soap to be in parity with https://github.com/Zeto-Ltd/clj-soap
[Card](https://rentpath.atlassian.net/browse/SRV-3150)

The rentpath clj-soap app will more than likely be used in the new sfmc-sink app. The rentpath/clj-soap app is way out of date. The Zeto-Ltd/clj-soap lib is more up to date, but there is no repository jar for it. The Zeto-Ltd lib is also loosely based off of the rentpath one as well as being basically only two relatively small namespaces. It is eaiser to just update the rentpath version.